libgo: merge Go 1.21 stdlib+runtime (WIP: conflicts pending)

+++ libgo/go/time/sys_unix.go
@@ -8,12 +8,20 @@ package time
 
 import (
 	"errors"
+	"runtime"
 	"syscall"
 )
 
 // for testing: whatever interrupts a sleep
 func interrupt() {
-	syscall.Kill(syscall.Getpid(), syscall.SIGCHLD)
+	// There is no mechanism in wasi to interrupt the call to poll_oneoff
+	// used to implement runtime.usleep so this function does nothing, which
+	// somewhat defeats the purpose of TestSleep but we are still better off
+	// validating that time elapses when the process calls time.Sleep than
+	// skipping the test altogether.
+	if runtime.GOOS != "wasip1" {
+		syscall.Kill(syscall.Getpid(), syscall.SIGCHLD)
+	}
 }
 
 func open(name string) (uintptr, error) {
